Consolidate PKI choices in a single file
It is easier to evaluate curves, curve parameters, signature algorithms,
key lengths, certificate constraints, and validity periods when they are
all in one place.
