+++ internal/pki/common.go
@@ -20,6 +20,7 @@ import (
 	"crypto/elliptic"
 	"crypto/rand"
 	"crypto/x509"
+	"crypto/x509/pkix"
 	"math/big"
 	"time"
 )
@@ -43,4 +44,63 @@ func generateKey() (*ecdsa.PrivateKey, error) {
 // generateSerialNumber returns a random 128-bit integer.
 func generateSerialNumber() (*big.Int, error) {
 	return rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), 128))
+}
+
+func generateLeafCertificate(
+	signer *x509.Certificate, signerPrivate *ecdsa.PrivateKey,
+	signeePublic *ecdsa.PublicKey, serialNumber *big.Int,
+	commonName string, dnsNames []string,
+) (*x509.Certificate, error) {
+	const leafExpiration = time.Hour * 24 * 365
+	const leafStartValid = time.Hour * -1
+
+	now := currentTime()
+	template := &x509.Certificate{
+		BasicConstraintsValid: true,
+		DNSNames:              dnsNames,
+		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment,
+		NotBefore:             now.Add(leafStartValid),
+		NotAfter:              now.Add(leafExpiration),
+		SerialNumber:          serialNumber,
+		SignatureAlgorithm:    certificateSignatureAlgorithm,
+		Subject: pkix.Name{
+			CommonName: commonName,
+		},
+	}
+
+	bytes, err := x509.CreateCertificate(rand.Reader, template, signer,
+		signeePublic, signerPrivate)
+
+	parsed, _ := x509.ParseCertificate(bytes)
+	return parsed, err
+}
+
+func generateRootCertificate(
+	privateKey *ecdsa.PrivateKey, serialNumber *big.Int,
+) (*x509.Certificate, error) {
+	const rootCommonName = "postgres-operator-ca"
+	const rootExpiration = time.Hour * 24 * 365 * 10
+	const rootStartValid = time.Hour * -1
+
+	now := currentTime()
+	template := &x509.Certificate{
+		BasicConstraintsValid: true,
+		IsCA:                  true,
+		KeyUsage:              x509.KeyUsageCertSign | x509.KeyUsageCRLSign,
+		MaxPathLenZero:        true, // there are no intermediate certificates
+		NotBefore:             now.Add(rootStartValid),
+		NotAfter:              now.Add(rootExpiration),
+		SerialNumber:          serialNumber,
+		SignatureAlgorithm:    certificateSignatureAlgorithm,
+		Subject: pkix.Name{
+			CommonName: rootCommonName,
+		},
+	}
+
+	// A root certificate is self-signed, so pass in the template twice.
+	bytes, err := x509.CreateCertificate(rand.Reader, template, template,
+		privateKey.Public(), privateKey)
+
+	parsed, _ := x509.ParseCertificate(bytes)
+	return parsed, err
 }
