+++ internal/pki/leaf.go
@@ -17,20 +17,7 @@ package pki
 
 import (
 	"context"
-	"crypto/ecdsa"
-	"crypto/rand"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"math/big"
-	"time"
 )
-
-// LeafCertificate is a certificate and private key pair that can be validated
-// by RootCertificateAuthority.
-type LeafCertificate struct {
-	Certificate Certificate
-	PrivateKey  PrivateKey
-}
 
 // LeafCertIsBad checks at least one leaf cert has been generated, the basic constraints
 // are valid and it has been verified with the root certpool
@@ -44,33 +31,4 @@ func LeafCertIsBad(
 	namespace string,
 ) bool {
 	return !rootCertCA.leafIsValid(leaf)
-}
-
-func generateLeafCertificate(
-	signer *x509.Certificate, signerPrivate *ecdsa.PrivateKey,
-	signeePublic *ecdsa.PublicKey, serialNumber *big.Int,
-	commonName string, dnsNames []string,
-) (*x509.Certificate, error) {
-	const leafExpiration = time.Hour * 24 * 365
-	const leafStartValid = time.Hour * -1
-
-	now := currentTime()
-	template := &x509.Certificate{
-		BasicConstraintsValid: true,
-		DNSNames:              dnsNames,
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment,
-		NotBefore:             now.Add(leafStartValid),
-		NotAfter:              now.Add(leafExpiration),
-		SerialNumber:          serialNumber,
-		SignatureAlgorithm:    certificateSignatureAlgorithm,
-		Subject: pkix.Name{
-			CommonName: commonName,
-		},
-	}
-
-	bytes, err := x509.CreateCertificate(rand.Reader, template, signer,
-		signeePublic, signerPrivate)
-
-	parsed, _ := x509.ParseCertificate(bytes)
-	return parsed, err
 }
