+++ internal/pki/root.go
@@ -15,40 +15,6 @@
 
 package pki
 
-import (
-	"crypto/ecdsa"
-	"crypto/rand"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"math/big"
-	"time"
-)
-
-// RootCertificateAuthority is a certificate and private key pair that can
-// generate other certificates.
-type RootCertificateAuthority struct {
-	Certificate Certificate
-	PrivateKey  PrivateKey
-}
-
-// NewRootCertificateAuthority generates a new key and self-signed certificate
-// for issuing other certificates.
-func NewRootCertificateAuthority() (*RootCertificateAuthority, error) {
-	var root RootCertificateAuthority
-	var serial *big.Int
-
-	key, err := generateKey()
-	if err == nil {
-		serial, err = generateSerialNumber()
-	}
-	if err == nil {
-		root.PrivateKey.ecdsa = key
-		root.Certificate.x509, err = generateRootCertificate(key, serial)
-	}
-
-	return &root, err
-}
-
 // RootCAIsBad checks that at least one root CA has been generated and that
 // all returned certs are CAs and not expired
 //
@@ -58,34 +24,4 @@ func NewRootCertificateAuthority() (*RootCertificateAuthority, error) {
 // certificate for our cert if it is the one that fails.
 func RootCAIsBad(root *RootCertificateAuthority) bool {
 	return !RootIsValid(root)
-}
-
-func generateRootCertificate(
-	privateKey *ecdsa.PrivateKey, serialNumber *big.Int,
-) (*x509.Certificate, error) {
-	const rootCommonName = "postgres-operator-ca"
-	const rootExpiration = time.Hour * 24 * 365 * 10
-	const rootStartValid = time.Hour * -1
-
-	now := currentTime()
-	template := &x509.Certificate{
-		BasicConstraintsValid: true,
-		IsCA:                  true,
-		KeyUsage:              x509.KeyUsageCertSign | x509.KeyUsageCRLSign,
-		MaxPathLenZero:        true, // there are no intermediate certificates
-		NotBefore:             now.Add(rootStartValid),
-		NotAfter:              now.Add(rootExpiration),
-		SerialNumber:          serialNumber,
-		SignatureAlgorithm:    certificateSignatureAlgorithm,
-		Subject: pkix.Name{
-			CommonName: rootCommonName,
-		},
-	}
-
-	// A root certificate is self-signed, so pass in the template twice.
-	bytes, err := x509.CreateCertificate(rand.Reader, template, template,
-		privateKey.Public(), privateKey)
-
-	parsed, _ := x509.ParseCertificate(bytes)
-	return parsed, err
 }
